Rename: Button, MesssageChip 컴포넌트 디렉터리 재구성
