--- src/components/shared/Button/type.ts
@@ -1,11 +1,0 @@
-export interface ButtonStyledType {
-	width?: string;
-	bgColor?: 'primary' | 'normal';
-	fontWeight: 'thin' | 'light' | 'normal' | 'medium' | 'semiBold' | 'bold';
-}
-
-export interface ButtonType extends ButtonStyledType {
-	type: 'button' | 'submit';
-	onClick?: () => void;
-	children: React.ReactNode;
-}
