+++ src/components/shared/Chip/MessageChip/MessageChip.type.ts
@@ -1,0 +1,3 @@
+export type Props = {
+	message: string;
+};
