--- src/components/shared/Chip/MessageChip/type.ts
@@ -1,3 +1,0 @@
-export interface MessageChipViewType {
-	message: string;
-}
